New set of changes based on the MuleSoft checklist
1) Updated the Mule Runtime to 4.3.0
2) Moved the Operations classes to the internal package
3) That meant I could move classes back to the internal package (sorry)
4) Changed the Category to CERTIFIED
5) Added LICENSE.MD
6) Added an example line for logging MarkLogic in the example project. Also added a couple of logging lines.
