+++ examples/demo-project/pom.xml
@@ -60,7 +60,7 @@
     <dependency>
       <groupId>com.marklogic</groupId>
       <artifactId>marklogic-mule-connector</artifactId>
-      <version>2.0.0-SNAPSHOT</version>
+      <version>2.0.0</version>
       <classifier>mule-plugin</classifier>
     </dependency>
     <dependency>
