+++ src/main/java/com/marklogic/mule/connector/internal/operation/Operations.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.connector.api;
+package com.marklogic.mule.connector.internal.operation;
 
 import com.marklogic.client.DatabaseClient;
 import com.marklogic.mule.connector.internal.error.provider.ExecuteErrorsProvider;
