+++ src/main/java/com/marklogic/mule/connector/internal/provider/AuthenticationType.java
@@ -13,23 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.connector.api.provider;
+package com.marklogic.mule.connector.internal.provider;
 
-import com.marklogic.client.DatabaseClient;
+public enum AuthenticationType {
 
-public enum ConnectionType {
-    DIRECT {
-        @Override
-        public DatabaseClient.ConnectionType getMarkLogicConnectionType() {
-            return DatabaseClient.ConnectionType.DIRECT;
-        }
-    },
-    GATEWAY {
-        @Override
-        public DatabaseClient.ConnectionType getMarkLogicConnectionType() {
-            return DatabaseClient.ConnectionType.GATEWAY;
-        }
-    };
+    DIGEST,
+    BASIC,
+    MARKLOGIC_CLOUD,
+    CERTIFICATE,
+    KERBEROS,
+    SAML;
 
-    public abstract DatabaseClient.ConnectionType getMarkLogicConnectionType();
 }
