+++ src/main/java/com/marklogic/mule/connector/internal/provider/QueryFormat.java
@@ -13,15 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.connector.api.provider;
+package com.marklogic.mule.connector.internal.provider;
 
-public enum AuthenticationType {
-
-    DIGEST,
-    BASIC,
-    MARKLOGIC_CLOUD,
-    CERTIFICATE,
-    KERBEROS,
-    SAML;
-
+public enum QueryFormat {
+    JSON,
+    XML;
 }
