+++ src/test/java/com/marklogic/mule/extension/SearchDocumentsWithMetadataTest.java
@@ -1,6 +1,6 @@
 package com.marklogic.mule.extension;
 
-import com.marklogic.mule.connector.api.provider.DocumentAttributes;
+import com.marklogic.mule.connector.internal.provider.DocumentAttributes;
 import org.junit.Test;
 
 import java.util.List;
