+++ src/test/java/com/marklogic/mule/extension/WriteDocumentTest.java
@@ -1,7 +1,7 @@
 package com.marklogic.mule.extension;
 
 
-import com.marklogic.mule.connector.api.provider.DocumentAttributes;
+import com.marklogic.mule.connector.internal.provider.DocumentAttributes;
 import org.junit.Test;
 
 import java.util.HashSet;
